support commonJS modules, via webpack
