+++ index.html
@@ -105,7 +105,7 @@
   </script>
 
 
-  <script src="src/color-thief.js"></script>
+  <script src="dist/color-thief.js"></script>
   <script src="examples/js/jquery.js"></script>
   <script src="examples/js/mustache.js"></script>
   <script src="examples/js/demo.js"></script>
